+++ src/spine/structure/time/schedule.hpp
@@ -17,7 +17,7 @@ public:
     struct Block {
         k_time_s start; // start time of block
         k_time_s duration; // length of block
-        double value; // active value for block
+        float value; // active value for block
     };
 
     struct Config {
@@ -30,7 +30,7 @@ public:
     Schedule(const std::initializer_list<Block>& blocks);
 
     /// Returns the value in the block of `t` or 0
-    double value_at(const k_time_s t) const;
+    float value_at(const k_time_s t) const;
 
     /// Returns the start of the block after time `t` or the maximum length of 24 hours
     k_time_s start_of_next_block(const k_time_s t) const;
